Include delay until data is expected to be ready
Without this delay the host is sending continuous read requests to the sensor. However the sensor is busy producing the data and does not respond to those requests. Observations with an oscilloscope indicate that each request deteriorates the SDA and SCL and potentially affects the sensor's and potentially the hosts' life time. SCA and SCL signal high level decreases and fluctuates when 0x40(R) is produced, indicating that hosts can not maintain a stable signal level on the i2c bus.